+++ packages/@wdio_electron-utils/src/binaryPath.ts
