+++ packages/@wdio_electron-utils/src/index.ts
@@ -1,3 +1,3 @@
-export { getElectronVersion } from './getElectronVersion';
-export { getAppBuildInfo } from './getAppBuildInfo';
-export { getBinaryPath } from './getBinaryPath';
+export { getElectronVersion } from './electronVersion';
+export { getAppBuildInfo } from './appBuildInfo';
+export { getBinaryPath } from './binaryPath';
