+++ packages/@wdio_electron-utils/test/binaryPath.spec.ts
